Fixed serious race condition exploits

+++ libs/libc/alloc.c
@@ -40,7 +40,7 @@ static int liballoc_free(void *ptr, u32 p)
 
 #endif
 
-static u8 locked = 0;
+static u32 locked = 0;
 
 static int liballoc_lock(void)
 {
