+++ libs/libc/inc/cpu.h
@@ -14,7 +14,7 @@ void outb(u16 port, u8 data);
 void outw(u16 port, u16 data);
 void outl(u16 port, u32 data);
 
-static inline void spinlock(u8 *ptr)
+static inline void spinlock(u32 *ptr)
 {
 	u32 prev;
 	do
